+++ web/src/global/GlobalAlerts.tsx
@@ -2,6 +2,7 @@ import { parseISO } from 'date-fns'
 import differenceInDays from 'date-fns/differenceInDays'
 import * as React from 'react'
 import { Subscription } from 'rxjs'
+import * as semver from 'semver'
 import { Markdown } from '../../../shared/src/components/Markdown'
 import { isSettingsValid, SettingsCascadeProps } from '../../../shared/src/settings/settings'
 import { renderMarkdown } from '../../../shared/src/util/markdown'
@@ -57,7 +58,13 @@ export class GlobalAlerts extends React.PureComponent<Props, State> {
                         {this.props.isSiteAdmin &&
                             this.state.siteFlags.updateCheck &&
                             !this.state.siteFlags.updateCheck.errorMessage &&
-                            this.state.siteFlags.updateCheck.updateVersionAvailable && (
+                            this.state.siteFlags.updateCheck.updateVersionAvailable &&
+                            ((isSettingsValid<Settings>(this.props.settingsCascade) &&
+                                this.props.settingsCascade.final['alerts.showPatchUpdates'] !== false) ||
+                                isMinorUpdateAvailable(
+                                    this.state.siteFlags.productVersion,
+                                    this.state.siteFlags.updateCheck.updateVersionAvailable
+                                )) && (
                                 <UpdateAvailableAlert
                                     className="global-alerts__alert"
                                     updateVersionAvailable={this.state.siteFlags.updateCheck.updateVersionAvailable}
@@ -113,4 +120,16 @@ export class GlobalAlerts extends React.PureComponent<Props, State> {
             </div>
         )
     }
+}
+
+function isMinorUpdateAvailable(currentVersion: string, updateVersion: string): boolean {
+    const cv = semver.parse(currentVersion, { loose: false })
+    const uv = semver.parse(updateVersion, { loose: false })
+    // If either current or update versions aren't semvers (e.g., a user is on a date-based build version, or "dev"),
+    // always return true and allow any alerts to be shown. This has the effect of simply deferring to the response
+    // from Sourcegraph.com about whether an update alert is needed.
+    if (cv === null || uv === null) {
+        return true
+    }
+    return semver.major(cv) !== semver.major(uv) || semver.minor(cv) !== semver.minor(uv)
 }